requirements.txt
fastapi
uvicorn
httpx
msal
python-dotenv
openai
pydantic